Update patch set 1
Patch Set 1:

Hi Chris,

We have one question on the Proxy setting in the base-r dockerfile,

What in heavens is this HARDCODED IP address? If you need a proxy it should be an environment variable
- is this proxy variable setting from Jenkins configuration? Or Onboarding code as we are doing currently 

We are correcting the remaining review comments.

Patch-set: 1
Reviewer: Gerrit User 15 <15@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0