Update patch set 4
Patch Set 4:

Not sure that 4.0.1 in pom file is the best value. My concern here is that we have another merged change for Demeter (6255) with 4.0.0 in pom file. I think we must have the same version in pom file for all the change in the same build am I right or wrong ?

Patch-set: 4