Update patch set 2
Patch Set 2:

First we need to revert the old failed change, which I have submitted in this change set.

Then we will start new.  You can always make sure a change is at tip of master by using the REBASE button at the top-right of the Gerrit UI.  That rebases onto master.

I'm also checking the merge job for the onboarding-base-r image to make sure *that* is working correctly.  That's obviously a prerequisite here.

Patch-set: 2