Update patch set 1
Patch Set 1: Code-Review-1

I  don't understand why we have ACUMOS-4109 in the commit message, this Jira ticket is related to a problem with docker tag command. In fact there was a mistake in the comand so no code change for on-boarding.

Then could you tell me to which defect or US the code change in OnboardingController.java is related for   

thanks

Patch-set: 1
Reviewer: Gerrit User 212 <212@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=-1